backend/pyproject.toml
[project]
name = "rent-manager-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "reportlab>=4.2.0",
  "supabase>=2.5.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]
